+++ Logibooks.Core/Controllers/RegistersController.cs
@@ -320,6 +320,12 @@ public class RegistersController(
             return _404Register(id);
         }
 
+        bool hasOrders = await _db.Orders.AnyAsync(r => r.RegisterId == id);
+        if (hasOrders)
+        {
+            return _409Register();
+        }
+
         _db.Registers.Remove(register);
         await _db.SaveChangesAsync();
         _logger.LogDebug("DeleteRegister returning '204 No content'");
